Generate Android SDK 25 image
This change adds a new target to the makefile to generate

     android/images/api-25-alpha/Dockerfile

which is a Dockerfile that generates a CircleCI blessed image for
running Android builds. I have alpha in the tag name so that we can
iterate on the container image.

I have punted on generating other Android API levels for now, until we
can see how this is working with customers.