Only edit the real mongo entrypoint file
In mongo:3.6, there is only a file at /usr/local/bin/docker-entrypoint.sh.
In earlier versions, there is also a symlink at /entrypoint.sh.

Previously, the attempt to edit the symlink without sed's --follow-symlinks
option was causing the symlink to be rewritten as a real file.

By now ignoring the symlink and only editing the real file. both 3.6 and
pre-3.6 versions are cleanly supported.