Add pre-deployment testing
After some isolated testing in my sandbox project to ensure this works,
I'm taking another pass at this.

To run goss tests successfully, we need to modify (in this case, add) the entrypoint for each image.

To do that, we make a copy of the Dockerfile in question, `cat` a few extra lines to this copy that include a modified entrypoint (it runs `sleep 600`, essentially), then build a new test image with the modified Dockerfile, and run goss tests on the test image.