Improve experience/function for community contributors
Currently, a bunch of logic in this repo prevents open-source contributions from being particularly helpful—i.e., we pull in a bunch of other GitHub repos that community members won't have access to, and we push stuff to a bunch of Docker Hub orgs that are limited to CircleCI employees.

We do this for good reason, but we should provide a way for open-source contributions to build and test, to the extent possible. This isn't an optimal solution, but it works:

Use forked PR-only environment variables to set up some control flow. Halt the circleci-dockerfiles and example-images steps if we're on a forked PR. Similarly, build and test—but don't push—images if we're on a forked PR.

Changes are in build.sh and config.yml.